requirements.txt
joblib==1.4.2
numpy==2.3.0
pandas==2.3.0
scikit_fuzzy==0.5.0
scikit_learn==1.4.2
streamlit==1.35.0
tensorflow_cpu==2.19.0
